service/pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "anthropic>=0.67.0",
    "databases[postgresql,sqlite]>=0.9.0",
    "fastapi[standard]>=0.116.1",
    "fastmcp>=2.10.5",
    "httpx>=0.28.1",
    "mcp>=1.11.0",
    "openai>=1.98.0",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic-graph>=0.4.10",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.4",
    "sqlmodel>=0.0.24",
    "types-requests>=2.32.4.20250611",
    "websockets>=13.0,<14.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
]

[tool.black]
line-length = 119
target-version = ['py312']
include = '.*\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations/versions
)/
'''

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
src_paths = ["."]

[tool.mypy]
python_version = "3.13"
disallow_untyped_decorators = false
ignore_missing_imports = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
mypy_path = "."
namespace_packages = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true
exclude = [
    ".*/migrations/.*",
    "migrations/.*",
    ".*/__pycache__/.*",
    ".*/\\.mypy_cache/.*",
    "\\.venv/.*"
]

[tool.pylance]
extraPaths = ["."]
autoImportCompletions = true
autoSearchPaths = true

[tool.pyright]
include = ["."]
exclude = ["**/__pycache__", "**/.mypy_cache", "*/migrations/*", "migrations/"]
extraPaths = ["."]
pythonVersion = "3.13"
typeCheckingMode = "basic"

# 针对特定模块的配置
[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true
